* pulled apart and organized all the database connectors, validators and writers into their own separate files for maintainability * cleaned up imports and comments in all examples
